# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IoT 全栈开发工作室的招新管理后台 API。\n统一响应体是 {code, message, data}；除公开端点（注册/登录/发码/刷新/登出）外都需要 Authorization: Bearer \u003caccess_token\u003e。",
        "title": "iot-pillot API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "description": "通过用户名**或邮箱** + 密码登录，返回双令牌（access_token + refresh_token）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回双令牌",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.LoginResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误（不区分二者）",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "description": "退出登录接口（需要 access token 和 refresh token）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "用户退出登录接口",
                "parameters": [
                    {
                        "description": "退出登录请求参数（refresh_token）",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefreshReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "退出登录成功",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "角色由服务端**现查 Redis 管理员名单**得到，而不是回显令牌里的快照，",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "获取当前登录用户",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证、令牌失效或账号已不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "只改令牌本人的资料，body 不接受 user_id 等目标参数（未知键直接忽略）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "更新本人资料",
                "parameters": [
                    {
                        "description": "资料字段补丁",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateMePatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的完整资料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "401": {
                        "description": "未认证、令牌失效或账号已不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "description": "使用 refresh token 获取新的一对令牌，旧令牌立即失效（单端登录）。不返回用户 ID。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefreshReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回新的双令牌",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.RefreshResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "401": {
                        "description": "refresh token 无效或已失效",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "注册新用户，返回双令牌。邮箱唯一；必须先通过 /send-verify-code 取得验证码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回双令牌",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.LoginResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误、邮箱已注册或验证码错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/send-verify-code": {
            "post": {
                "description": "发送邮箱验证码，5 分钟内有效、只能使用一次、猜错 5 次即作废。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "发送验证码接口",
                "parameters": [
                    {
                        "description": "验证目标",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.VerifyCodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发送验证码成功",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "429": {
                        "description": "请求过于频繁",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "request.RefreshReq": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "request.RegisterReq": {
            "type": "object",
            "required": [
                "code",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                },
                "email": {
                    "type": "string",
                    "maxLength": 50
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
...
        "request.UpdateMePatch": {
            "type": "object",
            "properties": {
                "class": {
                    "description": "班级；null 清除，缺省不动",
                    "type": "string",
                    "example": "物联网工程 2301"
                },
                "direction": {
                    "description": "方向；null/空串清除，缺省不动",
                    "type": "string",
                    "example": "front-end"
                },
                "qq": {
                    "description": "QQ；null 清除，缺省不动",
                    "type": "string",
                    "example": "1044696157"
                },
                "student_id": {
                    "description": "学号；null 清除（0 视为未填），缺省不动",
                    "type": "integer"
                }
            }
        },
        "request.VerifyCodeReq": {
            "type": "object",
            "required": [
                "verifier",
                "verifier_type"
            ],
            "properties": {
                "verifier": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                },
                "verifier_type": {
                    "type": "string",
                    "enum": [
                        "email",
                        "phone"
                    ]
                }
            }
        },
...
        "response.LoginResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.MeResp": {
            "type": "object",
            "properties": {
                "detail": {
                    "$ref": "#/definitions/response.UserProfileDetail"
                },
                "name": {
                    "description": "Name 是姓名（domain.User.Name），与 username 同源但语义独立：\n前端个人设置页把它当只读展示项。",
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.RefreshResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "response.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码：0 成功，非 0 失败",
                    "type": "integer"
                },
                "data": {
                    "description": "业务数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "response.UserProfileDetail": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "qq": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "登录得到的 access_token。Swagger UI 的 Authorize 里只填 token 本身，前缀由 UI 补。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}